Made deployment more reliable

+++ deployment/mock-data.sh
@@ -13,12 +13,12 @@ function redirect_dns {
 
 function create_mockdata {
     mkdir -p /var/www/html/api/3.4
-    rm /var/www/html/api/3.4/forecast.php
+    test -e /var/www/html/api/3.4/forecast.php && rm /var/www/html/api/3.4/forecast.php
     ln -s /vagrant/deployment/mock-data/forecast.json /var/www/html/api/3.4/forecast.php
 }
 
 function install_config {
-    rm /etc/apache2/sites-available/mock-data.conf
+    test -e /etc/apache2/sites-available/mock-data.conf && rm /etc/apache2/sites-available/mock-data.conf
     ln -s /vagrant/deployment/mock-data/mock-data.conf /etc/apache2/sites-available/
     a2ensite mock-data
 }
